Moved card tag to its component
