+++ src/components/MovieCard/MovieCard.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import './MovieCard.scss';
 
 export const MovieCard = ({ title, description, imgUrl, imdbUrl }) => (
-  <>
+  <div className="card">
     <div className="card-image">
       <figure className="image is-4by3">
         <img
@@ -33,7 +33,7 @@ export const MovieCard = ({ title, description, imgUrl, imdbUrl }) => (
         <a href={imdbUrl}>IMDB</a>
       </div>
     </div>
-  </>
+  </div>
 );
 
 MovieCard.propTypes = {
