+++ src/components/MoviesList.jsx
@@ -2,14 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import './MoviesList.scss';
-import { MovieCard } from '../MovieCard';
+import { MovieCard } from './MovieCard';
 
 export const MoviesList = ({ movies }) => (
   <div className="movies">
     {movies.map(movie => (
-      <div className="card" key={movie.imdbId}>
-        <MovieCard {...movie} />
-      </div>
+      <MovieCard {...movie} key={movie.imdbId} />
     ))}
   </div>
 );
