+++ src/components/MoviesList.scss
